// web/lib/api/generated/models/itemTemplateResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * GroceryManager.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { TrackingUnit } from './trackingUnit';

export interface ItemTemplateResponse {
  id: string;
  name: string;
  defaultCategoryKey: string;
  defaultTrackingUnit: TrackingUnit;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  sortOrder: number | string;
}